pyproject.toml
[tool.poetry]
name = "beanstalk"
version = "0.1.0"
description = "aaaaaaa"
authors = ["Cassidy Dingenskirchen <admin@15318.de>"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "~1.6"
gTTS = "^2.2.2"
urllib3 = "^1.26.4"
aiohttp = "^3.7.4"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"

[tool.poetry.scripts]
blimp = "beanstalk.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
